Replace timing test with crash-reproducing regression test
The previous "runs after commit stack" test only pinned the microtask
mechanism, not the property we care about. Replace it with an integration
test that renders a layout-feedback component and asserts render does not
throw "Maximum update depth exceeded" — this fails without the fix and is
agnostic to how the deferral is implemented. Keep the removed-listener test,
which guards the stale-listener edge case the deferral introduces.
